refactor features etc
